commit siete ya se puede envier correo y velos rango de los precios, pero no sumo el email.properties por seguridad si necesita perdilo

+++ src/main/java/com/example/demo/DTOs/ProductoDTO/RangoDePreciosDTO.java
@@ -2,10 +2,12 @@ package com.example.demo.DTOs.ProductoDTO;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class RangoDePreciosDTO {
-    private Double precioMinimo;
-    private Double precioMaximo;
+    private BigDecimal precioMinimo;
+    private BigDecimal precioMaximo;
 }
